requirements.txt
Flask
python-dotenv
flask-cors
Pillow
Werkzeug
gunicorn
flask_limiter
Blueprint
